requirements.txt
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu118
flask==2.3.2
gunicorn==20.1.0
sox==1.4.1
ffmpeg-python==0.2.0
git+https://github.com/isi-nlp/uroman.git
halo==0.0.31
omegaconf==2.0.6
hydra-core==1.0.7
transformers[torch]==4.31.0
python-dotenv
pydub==0.25.1
flask-cors
numpy==1.26.4
